Moved workspace-menu
