+++ ui/packages/catalog-ui-search/src/main/webapp/extension-points/routes/definitions/base/openWorkspace.js
@@ -18,7 +18,7 @@
     patterns: ['workspaces/:id'],
     component: 'component/content/content.view',
     menu: {
-      component: 'react-component/container/workspace-menu/workspace-menu',
+      component: 'react-component/workspace-menu/container',
     },
   },
 })
