+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/workspace-menu/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './workspace-menu'
+export { default } from './container'
